K fix

+++ src/main/java/SF/Filter/FilterFunction.java
@@ -3,5 +3,5 @@ package SF.Filter;
 public interface FilterFunction
 {
     Double getValue(final int i);
-    int getK(final Double samplingFrequency, final Double frequency);
+    Double getK(final Double samplingFrequency, final Double frequency);
 }
